Handle spurious prompts when configuring packages.
The prompts arose in part  (ninjasphere-factory-test) because fpm changed its behaviour to mark all
files in /etc/ as configuration files by default when previously
they were treated as normal files. The new package then gets blocked
because of the existence of a file in a place which is now
a configuration location.

Signed-off-by: Jon Seymour <jon@wildducktheories.com>